requirements.txt
astroid==2.12.13
attrs==22.2.0
bcrypt==4.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
dill==0.3.6
flake8==6.0.0
Flask==2.2.3
Flask-Cors==3.0.10
hypothesis==6.61.0
idna==3.4
iniconfig==1.1.1
isort==5.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.1
mccabe==0.7.0
packaging==22.0
platformdirs==2.6.2
plotly==5.8.2
pluggy==1.0.0
pycodestyle==2.10.0
pyflakes==3.0.1
pygame==2.1.3.dev8
Pygments==2.14.0
PyJWT==2.6.0
pylint==2.15.9
pytest==7.2.0
python-dotenv==1.0.0
python-ta==2.4.2
requests==2.28.1
six==1.16.0
sortedcontainers==2.4.0
tabulate==0.8.10
tenacity==8.1.0
toml==0.10.2
tomlkit==0.11.6
typeguard==2.13.3
urllib3==1.26.13
Werkzeug==2.2.3
wrapt==1.14.1
